CLOUD-2670: Added test that reproduces the issue
CLOUD-2716: Added assert validating proper AIO usage
CLOUD-2716: Fixed undefined behavior when initializing MediatorConnector and MediatorStunClient
META-125: Coding style fixes
META-184: Added possibility to substitute analytics manifests
VMS-10372: Speedup server stopping during AV resource initialization
VMS-10862: Partially visible bookmark doesn't look hovered while cursor out of time slider
VMS-11183: Fixed incorrect 'read only' flag display when exporting video
VMS-11185: Code readability improvement
VMS-11185: Sound placeholder in export preview doesn't display right
VMS-11616 Fixed incorrect html rendering in a dialog with update notice
VMS-11915: Removed the code for old update system
VMS-12699: Fixed possible zero pointer usage
VMS-12734:pre-patch update of the licensing lists
VMS-12784: Suppressing reconnect dialog when we are in 'installing update' phase
VMS-12803: Another fix for uploading offline updates
VMS-12821: Raised priority of GET/POST HTTP tunnel over 'HTTP connection upgrade' tunnel.
VMS-12893: Fixed css embedding to EULA form
VMS-12893: Showing EULA from updateInformation.eula
VMS-12922: Fixed empty resolution of video frames from USB cameras
VMS-12922: Parse video resolution for plugin cameras(support h263 codec)
VMS-12925: Client will try to reconnect to mediaserver after successful update
VMS-12943: Handle double-click as usual instead of the ptz command for ptz-redirected cameras
VMS-13060: Plugin and group names are not selectable anymore in Analytics Event Widget combobox
VMS-13069: All checks for updates can be proxied through mediaservers
VMS-13114: Removed the activeAnalyticsEngines method from QnVirtualCameraResource public interface
VMS-13162: Display motion search region only on camera, selected in the right panel
VMS-13214: There is ability to copy motion-only record schedule for single-stream camera
VMS-13223: Adjust tests to the new API. More fixes to rest handlers and helpers
VMS-13223: Wrap updateStatus and updateInformation responses into RestData
VMS-13224: Remove a redundant syscall
VMS-13256: Add even more logs
VMS-13256: Add logs on every init/reinit of resource
VMS-13256: Add more logs to the god of logs
VMS-13256: Fix not reiniting in thirdPartyResource after stream URL update, add logs
VMS-13258: Fix unit tests
VMS-13258: Make downloader process files with arbitrary paths
VMS-13258: Merge with vms_4.0_server
VMS-13262: Fixed compiling
VMS-13262: Renamed in mediaserver.ini: enableApiServerControl to enableApiDebug
VMS-13294: 'Reconstruct Resolution' feature for the Entropix (hidden under ini-config)
VMS-13296: Fix System PTZ presets for DS-2DE3304W-DE
VMS-13296: Make system/native PTZ presets switch more robust
VMS-13297: Fix rotated panoramic cameras display
VMS-13307: Make camera info files parsing algorithm more robust
VMS-13308: Manual graft of necessary changes from the vms_4.0 branch
VMS-13308: Remove unused functions
VMS-13309: Add Cloud libraries to the blacklists of Server and Client distribution build scripts
VMS-13310: Add a default initializer to the FileInfoEx struct member
VMS-13310: Refactor removing empty-only directories
VMS-13310: Restore the altered remove-empty-dirs timeout to the default value
VMS-13312: Fixed updates in compatibility mode
VMS-13315: Manual graft of necessary changes from 3.2
VMS-13315: Revert the O_SYNC flag when opening a file in the unbuffered mode
VMS-13315: Substiture direct int64 assigning with memcpy. Add default vacuum timeout
VMS-13315: Substiture direct int64 read assignment with memcpy. Open nxdb file in the unbuffered mode
VMS-13318: Optimized status watching for embedded Avi streams
VMS-13324: Add toWebClientStandardViolatingUrl method to Url
VMS-13334: Fix playback video via cloud
VMS-13338: Revert setting O_SYNC flag for the unbuffered mode
VMS-13341: Fix server crash on LDAP authorization
VMS-13345: Remove support of the DVD/Blueray from the desktop client
VMS-13346: Stop an async http client before destroying from a non-AIO thread
VMS-13349: Fix camera reinitialization for some cases
VMS-13349: Typo fix
VMS-13350: Fixed update widget hanging in 'checking' state
VMS-13353: Improve assert message, renamings.
VMS-13353: add 'updateSequence' check same as for regular servers
VMS-13353: review minor
VMS-13353: vms_cloud_integration_tests/P2pMode/FtDisconnectSystemFromCloud.AfterDisconnect_NonCloudNonce/1 test crash
VMS-13358: Enabled CloudMerge.resulting_system_can_be_accessed_via_temporary_cloud_credentials test
VMS-13358: Fixed deadlock when merging non-cloud systems
VMS-13358: Minor code improvement
VMS-13358: Minor refactoring to help fix the issue
VMS-13358: More logs
VMS-13358: When merging systems, making sure that cloud user params are taken from the proper system
VMS-13368: Remove licence activation limitations on ARMs
VMS-13371: Add all libnx_clusterdb* libraries to Server's and Client's distributions blacklists.
VMS-13371: Add the rest of Cloud libraries in Client's and Server's distribution blacklists
VMS-13371: Fix the bug: revert libcloud_db_client.so in Client's distribution.
VMS-13371: Fix the bug: revert libcloud_db_client.so in Server's distribution.
VMS-13381: Recording is not enabled correctly for cameras with different schedule
VMS-13389: Added test that reproduces the problem
VMS-13389: Cloud user offline login works right after binding system to the Cloud
VMS-13390: Fix the findDownloads() function
VMS-13390: Make the downloader store workers map by file-name-only key
VMS-13390: Make the p2p downloader store metadata files in a fixed directory
VMS-13392: Do not use USB serial numbers for Windows as camera identifier
VMS-13410: resample source UTC PTS to the encoder PTS manually to improve encoder bitrate stability
VMS-8220: Make the SysV mediaserver starting script not loop in case of a wrong service command
VMS-8220: Remove init.d scripts from the deb package completely
VMS-9385: Remove round-robing for channels
WEB-949: Fullscreen mode of a camera doesn't work in Chrome
WEB-958: Double click doesn't work to exit fullscreen mode.
WEB-962: Webclient doesn't work at IE 11.

<==== Mercurial data ====>
Branch: master

+++ vms/libs/common/src/nx/vms/common/p2p/downloader/private/storage.cpp
@@ -13,15 +13,36 @@
 
 using namespace std::chrono;
 
+namespace nx::vms::common::p2p::downloader {
+
 namespace {
 
 static constexpr qint64 kDefaultChunkSize = 1024 * 1024;
 static const QString kMetadataSuffix = ".vmsdownload";
 static const milliseconds kCleanupPeriod = 5min;
 
+static FileMetadata prepareInitialFileMetadata(const FileInformation& fileInformation,
+    const QDir& defaultDownloadsDirectory)
+{
+    FileMetadata info = fileInformation;
+    if (info.chunkSize <= 0)
+        info.chunkSize = kDefaultChunkSize;
+
+    if (info.absoluteDirectoryPath.isEmpty())
+    {
+        info.fullFilePath = defaultDownloadsDirectory.absoluteFilePath(info.name);
+        info.absoluteDirectoryPath = QFileInfo(info.fullFilePath).absolutePath();
+    }
+    else
+    {
+        info.fullFilePath = QDir(info.absoluteDirectoryPath).absoluteFilePath(info.name);
+    }
+
+    return info;
+}
+
 } // namespace
 
-namespace nx::vms::common::p2p::downloader {
 
 QN_FUSION_DECLARE_FUNCTIONS(FileMetadata, (json))
 
@@ -29,6 +50,7 @@ Storage::Storage(const QDir& downloadsDirectory, QObject* parent):
     QObject(parent),
     m_downloadsDirectory(downloadsDirectory)
 {
+    QDir().mkpath(metadataDirectoryPath());
     /* Cleanup expired files every 5 mins. */
     const auto timer = new QTimer(this);
     connect(timer, &QTimer::timeout, this, &Storage::cleanupExpiredFiles);
@@ -45,22 +67,31 @@ QDir Storage::downloadsDirectory() const
     return m_downloadsDirectory;
 }
 
-QStringList Storage::files() const
-{
-    QnMutexLocker lock(&m_mutex);
-    return m_fileInformationByName.keys();
-}
-
 QString Storage::filePath(const QString& fileName) const
 {
-    return m_downloadsDirectory.absoluteFilePath(fileName);
+    QnMutexLocker lock(&m_mutex);
+    const auto it = m_fileInformationByName.find(FileInformation::keyFromFileName(fileName));
+    if (it == m_fileInformationByName.cend())
+        return QString();
+
+    return it.value().fullFilePath;
 }
 
-FileInformation Storage::fileInformation(
-    const QString& fileName) const
+QStringList Storage::files() const
+{
+    QStringList result;
+    {
+        QnMutexLocker lock(&m_mutex);
+        for (auto it = m_fileInformationByName.cbegin(); it != m_fileInformationByName.cend(); ++it)
+            result.append(it.value().fullFilePath);
+    }
+    return result;
+}
+
+FileInformation Storage::fileInformation(const QString& fileName) const
 {
     QnMutexLocker lock(&m_mutex);
-    return m_fileInformationByName.value(fileName);
+    return m_fileInformationByName.value(FileInformation::keyFromFileName(fileName));
 }
 
 ResultCode Storage::addFile(FileInformation fileInformation, bool updateTouchTime)
@@ -80,21 +111,14 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
 
     QnMutexLocker lock(&m_mutex);
 
-    if (m_fileInformationByName.contains(fileInformation.name))
+    if (m_fileInformationByName.contains(fileInformation.key()))
         return ResultCode::fileAlreadyExists;
 
-    FileMetadata info = fileInformation;
-    const auto path = filePath(fileInformation.name);
-
-    if (info.chunkSize <= 0)
-        info.chunkSize = kDefaultChunkSize;
-
-    QFile file(path);
-
-    if (!file.exists())
+    FileMetadata info = prepareInitialFileMetadata(fileInformation, m_downloadsDirectory);
+    if (!QFile(info.fullFilePath).exists())
         return ResultCode::fileDoesNotExist;
 
-    const auto md5 = calculateMd5(path);
+    const auto md5 = calculateMd5(info.fullFilePath);
     if (md5.isEmpty())
         return ResultCode::ioError;
 
@@ -103,7 +127,7 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
     else if (info.md5 != md5)
         return ResultCode::invalidChecksum;
 
-    const auto size = calculateFileSize(path);
+    const auto size = calculateFileSize(info.fullFilePath);
     if (size < 0)
         return ResultCode::ioError;
 
@@ -113,7 +137,7 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
         return ResultCode::invalidFileSize;
 
     const int chunkCount = calculateChunkCount(info.size, info.chunkSize);
-    info.chunkChecksums = calculateChecksums(path, info.chunkSize);
+    info.chunkChecksums = calculateChecksums(info.fullFilePath, info.chunkSize);
     if (info.chunkChecksums.size() != chunkCount)
         return ResultCode::ioError;
 
@@ -122,7 +146,7 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
     if (!saveMetadata(info))
         return ResultCode::ioError;
 
-    m_fileInformationByName.insert(fileInformation.name, info);
+    m_fileInformationByName.insert(fileInformation.key(), info);
 
     lock.unlock();
 
@@ -137,31 +161,25 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
 
     QnMutexLocker lock(&m_mutex);
 
-    if (m_fileInformationByName.contains(fileInformation.name))
+    if (m_fileInformationByName.contains(fileInformation.key()))
         return ResultCode::fileAlreadyExists;
 
-    FileMetadata info = fileInformation;
-    const auto path = filePath(fileInformation.name);
-
-    if (info.chunkSize <= 0)
-        info.chunkSize = kDefaultChunkSize;
-
-    const auto fileDir = QFileInfo(path).absolutePath();
-    if (!QDir(fileDir).exists())
+    FileMetadata info = prepareInitialFileMetadata(fileInformation, m_downloadsDirectory);
+    if (!QDir(info.absoluteDirectoryPath).exists())
     {
-        if (!QDir().mkpath(fileDir))
+        if (!QDir().mkpath(info.absoluteDirectoryPath))
             return ResultCode::ioError;
     }
 
-    if (!info.md5.isEmpty() && calculateMd5(path) == info.md5)
+    if (!info.md5.isEmpty() && calculateMd5(info.fullFilePath) == info.md5)
     {
         info.status = FileInformation::Status::downloaded;
-        info.size = calculateFileSize(path);
+        info.size = calculateFileSize(info.fullFilePath);
         if (info.size < 0)
             return ResultCode::ioError;
 
         const int chunkCount = calculateChunkCount(info.size, info.chunkSize);
-        info.chunkChecksums = calculateChecksums(path, info.chunkSize);
+        info.chunkChecksums = calculateChecksums(info.fullFilePath, info.chunkSize);
         if (info.chunkChecksums.size() != chunkCount)
             return ResultCode::ioError;
 
@@ -188,7 +206,7 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
             info.chunkChecksums.resize(chunkCount);
         }
 
-        ResultCode reserveStatus = reserveSpace(path, info.size >= 0 ? info.size : 0);
+        ResultCode reserveStatus = reserveSpace(info.fullFilePath, info.size >= 0 ? info.size : 0);
         if (reserveStatus != ResultCode::ok)
             return reserveStatus;
 
@@ -198,7 +216,7 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
     if (!saveMetadata(info))
         return ResultCode::ioError;
 
-    m_fileInformationByName.insert(fileInformation.name, info);
+    m_fileInformationByName.insert(fileInformation.key(), info);
 
     lock.unlock();
 
@@ -207,12 +225,12 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
     return ResultCode::ok;
 }
 
-ResultCode Storage::updateFileInformation(
-    const QString& fileName, qint64 size, const QByteArray& md5)
+ResultCode Storage::updateFileInformation(const QString& fileName, qint64 size,
+    const QByteArray& md5)
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(FileInformation::keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -229,7 +247,7 @@ ResultCode Storage::updateFileInformation(
         const int chunkCount = calculateChunkCount(size, it->chunkSize);
         it->downloadedChunks.resize(chunkCount);
         it->chunkChecksums.resize(chunkCount);
-        resizeFailed = !QFile::resize(filePath(fileName), size);
+        resizeFailed = !QFile::resize(it->fullFilePath, size);
 
         updated = true;
     }
@@ -272,7 +290,7 @@ ResultCode Storage::setChunkSize(const QString& fileName, qint64 chunkSize)
 
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(FileInformation::keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -307,12 +325,11 @@ ResultCode Storage::setChunkSize(const QString& fileName, qint64 chunkSize)
     return ResultCode::ok;
 }
 
-ResultCode Storage::readFileChunk(
-    const QString& fileName, int chunkIndex, QByteArray& buffer)
+ResultCode Storage::readFileChunk(const QString& fileName, int chunkIndex, QByteArray& buffer)
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(FileInformation::keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -322,7 +339,7 @@ ResultCode Storage::readFileChunk(
     if (!it->downloadedChunks.testBit(chunkIndex))
         return ResultCode::invalidChunkIndex;
 
-    QFile file(filePath(it->name));
+    QFile file(it->fullFilePath);
     if (!file.open(QFile::ReadOnly))
         return ResultCode::ioError;
 
@@ -337,12 +354,11 @@ ResultCode Storage::readFileChunk(
     return ResultCode::ok;
 }
 
-ResultCode Storage::writeFileChunk(
-    const QString& fileName, int chunkIndex, const QByteArray& buffer)
+ResultCode Storage::writeFileChunk(const QString& fileName, int chunkIndex, const QByteArray& buffer)
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(FileInformation::keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -352,7 +368,7 @@ ResultCode Storage::writeFileChunk(
     if (it->status == FileInformation::Status::downloaded)
         return ResultCode::ioError;
 
-    QFile file(filePath(it->name));
+    QFile file(it->fullFilePath);
     if (!file.open(QFile::ReadWrite)) //< ReadWrite because WriteOnly implies Truncate.
         return ResultCode::ioError;
 
@@ -400,47 +416,49 @@ ResultCode Storage::deleteFile(const QString& fileName, bool deleteData)
             return resultCode;
     }
 
-    emit fileDeleted(fileName);
+    emit fileDeleted(filePath(fileName));
     return ResultCode::ok;
 }
 
 ResultCode Storage::deleteFileInternal(const QString& fileName, bool deleteData)
 {
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(FileInformation::keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
-    const auto path = filePath(it->name);
-
-    if (deleteData && QFile::exists(path))
+    if (deleteData && QFile::exists(it->fullFilePath))
     {
-        if (!QFile::remove(path))
+        if (!QFile::remove(it->fullFilePath))
             return ResultCode::ioError;
     }
 
-    const auto metadataFileName = this->metadataFileName(path);
+    const auto metadataFileName = this->metadataFileName(it->fullFilePath);
     if (QFile::exists(metadataFileName))
     {
         if (!QFile::remove(metadataFileName))
             return ResultCode::ioError;
     }
 
-    QDir dir = QFileInfo(path).absoluteDir();
-    while (dir != m_downloadsDirectory)
+    // We won't delete custom user directories.
+    if (it->absoluteDirectoryPath.contains(m_downloadsDirectory.absolutePath()))
     {
-        if (dir.entryList(QDir::AllEntries | QDir::NoDotAndDotDot).isEmpty())
+        QDir dir = QFileInfo(it->absoluteDirectoryPath).absoluteDir();
+        while (dir != m_downloadsDirectory)
         {
-            const auto name = dir.dirName();
+            if (dir.entryList(QDir::AllEntries | QDir::NoDotAndDotDot).isEmpty())
+            {
+                const auto name = dir.dirName();
 
-            if (!dir.cdUp())
-                break;
+                if (!dir.cdUp())
+                    break;
 
-            if (!dir.rmdir(name))
+                if (!dir.rmdir(name))
+                    break;
+            }
+            else
+            {
                 break;
-        }
-        else
-        {
-            break;
+            }
         }
     }
 
@@ -453,7 +471,7 @@ QVector<QByteArray> Storage::getChunkChecksums(const QString& fileName)
 {
     QnMutexLocker lock(&m_mutex);
 
-    const auto& fileInfo = fileMetadata(fileName);
+    const auto& fileInfo = fileMetadata(FileInformation::keyFromFileName(fileName));
     if (!fileInfo.isValid())
         return QVector<QByteArray>();
 
@@ -465,7 +483,7 @@ ResultCode Storage::setChunkChecksums(
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(FileInformation::keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -494,7 +512,7 @@ ResultCode Storage::setChunkChecksums(
     if (it->size < 0)
         return ResultCode::ok;
 
-    auto actualChecksums = calculateChecksums(filePath(fileName), it->chunkSize);
+    auto actualChecksums = calculateChecksums(it->fullFilePath, it->chunkSize);
     if (actualChecksums.size() != it->downloadedChunks.size())
         actualChecksums.resize(it->downloadedChunks.size());
 
@@ -520,7 +538,7 @@ void Storage::cleanupExpiredFiles()
     for (const FileMetadata& data: m_fileInformationByName)
     {
         if (data.ttl > 0 && data.touchTime + data.ttl <= currentTime)
-            expiredFiles.insert(data.name);
+            expiredFiles.insert(data.fullFilePath);
     }
 
     QSet<QString> successfullyDeletedFiles;
@@ -549,7 +567,7 @@ void Storage::findDownloads(bool waitForFinished)
     m_findDownloadsWatcher.setFuture(QtConcurrent::run(
         [this]()
         {
-            findDownloadsRecursively(m_downloadsDirectory);
+            findDownloadsImpl();
             cleanupExpiredFiles();
         }));
 
@@ -560,29 +578,22 @@ void Storage::findDownloads(bool waitForFinished)
     }
 }
 
-void Storage::findDownloadsRecursively(const QDir& dir)
+void Storage::findDownloadsImpl()
 {
-    for (const auto& entry: dir.entryInfoList(QDir::Files | QDir::Dirs | QDir::NoDotAndDotDot))
+    for (const auto& entry: QDir(metadataDirectoryPath()).entryInfoList(QDir::Files))
     {
         auto fileName = entry.absoluteFilePath();
-        if (entry.isDir())
+        if (!fileName.endsWith(kMetadataSuffix))
+            continue;
+
+        fileName.truncate(fileName.size() - kMetadataSuffix.size());
         {
-            findDownloadsRecursively(QDir(fileName));
+            NX_MUTEX_LOCKER lock(&m_mutex);
+            if (m_fileInformationByName.contains(FileInformation::keyFromFileName(fileName)))
+                continue;
         }
-        else if (fileName.endsWith(kMetadataSuffix))
-        {
-            fileName.truncate(fileName.size() - kMetadataSuffix.size());
-            if (entry.isFile())
-            {
-                const auto& name = m_downloadsDirectory.relativeFilePath(fileName);
-                {
-                    NX_MUTEX_LOCKER lock(&m_mutex);
-                    if (m_fileInformationByName.contains(name))
-                        continue;
-                }
-                loadDownload(fileName);
-            }
-        }
+
+        loadDownload(fileName);
     }
 }
 
@@ -650,7 +661,7 @@ QVector<QByteArray> Storage::calculateChecksums(const QString& filePath, qint64 
 
 bool Storage::saveMetadata(const FileMetadata& fileInformation)
 {
-    const auto fileName = metadataFileName(filePath(fileInformation.name));
+    const auto fileName = metadataFileName(fileInformation.fullFilePath);
 
     QFile file(fileName);
     if (!file.open(QFile::WriteOnly))
@@ -696,6 +707,11 @@ FileMetadata Storage::loadMetadata(const QString& fileName)
     return fileInfo;
 }
 
+QString Storage::metadataDirectoryPath() const
+{
+    return m_downloadsDirectory.absoluteFilePath(kMetadataSuffix);
+}
+
 FileMetadata Storage::fileMetadata(const QString& fileName) const
 {
     return m_fileInformationByName.value(fileName);
@@ -733,9 +749,7 @@ void Storage::checkDownloadCompleted(FileMetadata& fileInfo)
             return;
     }
 
-    const auto path = filePath(fileInfo.name);
-
-    const auto md5 = calculateMd5(path);
+    const auto md5 = calculateMd5(fileInfo.fullFilePath);
     if (md5 != fileInfo.md5)
     {
         fileInfo.status = FileInformation::Status::corrupted;
@@ -743,7 +757,7 @@ void Storage::checkDownloadCompleted(FileMetadata& fileInfo)
     }
 
     fileInfo.status = FileInformation::Status::downloaded;
-    fileInfo.chunkChecksums = calculateChecksums(path, fileInfo.chunkSize);
+    fileInfo.chunkChecksums = calculateChecksums(fileInfo.fullFilePath, fileInfo.chunkSize);
 }
 
 ResultCode Storage::reserveSpace(const QString& fileName, const qint64 size)
@@ -765,10 +779,11 @@ ResultCode Storage::reserveSpace(const QString& fileName, const qint64 size)
 
 QString Storage::metadataFileName(const QString& fileName)
 {
+    NX_ASSERT(!fileName.isEmpty());
     if (fileName.isEmpty())
         return QString();
 
-    return fileName + kMetadataSuffix;
+    return QDir(metadataDirectoryPath()).absoluteFilePath(QFileInfo(fileName).fileName() + kMetadataSuffix);
 }
 
 qint64 Storage::calculateChunkSize(qint64 fileSize, int chunkIndex, qint64 chunkSize)
@@ -785,6 +800,7 @@ qint64 Storage::calculateChunkSize(qint64 fileSize, int chunkIndex, qint64 chunk
         : fileSize - chunkSize * (chunkCount - 1);
 }
 
-QN_FUSION_ADAPT_STRUCT_FUNCTIONS(FileMetadata, (json), FileInformation_Fields (chunkChecksums))
+QN_FUSION_ADAPT_STRUCT_FUNCTIONS(FileMetadata, (json), FileInformation_Fields \
+    (chunkChecksums)(fullFilePath))
 
 } // nx::vms::common::p2p::downloader
